Add 'eslint' config
