+++ next.config.js
@@ -1,8 +1,8 @@
 require("dotenv").config();
 
 module.exports = {
-  // ENV variables
-  env: {
-    GITHUB_PRIVATE_TOKEN: process.env.GITHUB_PRIVATE_TOKEN,
-  }
+	// ENV variables
+	env: {
+		GITHUB_PRIVATE_TOKEN: process.env.GITHUB_PRIVATE_TOKEN,
+	},
 };
