+++ package.json
@@ -15,6 +15,7 @@
     "react-dom": "16.13.1"
   },
   "devDependencies": {
+    "@workwithizzi/eslint-config-react": "^0.0.2",
     "dotenv": "^8.2.0"
   }
 }
